Fixed modal/popup background
Changed the background for the popup (.bubblebanner) that appears when waiting for other players that says "This is a premium game", it was yellow background with white text. Also changed the modal that asks if you want to send an email to notify your friend when inviting him/her, was white over white.